Updated sharing
