+++ build-system/compile/bundles.config.extensions.json
@@ -1039,6 +1039,14 @@
     }
   },
   {
+    "name": "amp-story-share-menu",
+    "version": "0.1",
+    "latestVersion": "0.1",
+    "options": {
+      "hasCss": true
+    }
+  },
+  {
     "name": "amp-story-shopping",
     "version": "0.1",
     "latestVersion": "0.1",
