+++ extensions/amp-story-share-menu/0.1/amp-story-share-menu.css
@@ -1,6 +1,6 @@
 
 
-@import './amp-story-shadow-reset.css';
+@import '../../amp-story/1.0/amp-story-shadow-reset.css';
 @import './amp-story-share.css';
 
 .i-amphtml-story-share-menu {
